mysite/requirements.txt
django==3.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.6.0
stripe==2.56.0
dj-stripe==2.4.3
mysqlclient==2.0.3
PyMySQL==1.0.2
